@@ -1,7 +1,4 @@
-﻿global using System.Collections;
-global using System.Text;
-
-BitArray bits1 = new(9);
+﻿BitArray bits1 = new(9);
 bits1.SetAll(true);
 bits1.Set(1, false);
 bits1[5] = false;
