@@ -9,25 +9,24 @@ while (enumerator.MoveNext())
     enumerator = enumerator.Current as IEnumerator<IEnumerator> ?? throw new InvalidOperationException();
 }
 
-
 void MusicTitles()
 {
     MusicTitles titles = new();
-    foreach (var title in titles)
+    foreach (string title in titles)
     {
         Console.WriteLine(title);
     }
     Console.WriteLine();
 
     Console.WriteLine("reverse");
-    foreach (var title in titles.Reverse())
+    foreach (string title in titles.Reverse())
     {
         Console.WriteLine(title);
     }
     Console.WriteLine();
 
     Console.WriteLine("subset");
-    foreach (var title in titles.Subset(2, 2))
+    foreach (string title in titles.Subset(2, 2))
     {
         Console.WriteLine(title);
     }
